+++ next.config.js
@@ -87,4 +87,102 @@ module.exports = {
       },
     ];
   },
+  async redirects() {
+    return [
+      {
+        source: "/en",
+        has: [
+          {
+            type: "cookie",
+            key: "lang",
+            value: "fr",
+          },
+        ],
+        locale: false,
+        permanent: false,
+        destination: "/fr/accueil",
+      },
+      {
+        source: "/en",
+        has: [
+          {
+            type: "cookie",
+            key: "lang",
+            value: "en",
+          },
+        ],
+        locale: false,
+        permanent: false,
+        destination: "/home",
+      },
+      {
+        source: "/en/:path",
+        has: [
+          {
+            type: "cookie",
+            key: "lang",
+            value: "fr",
+          },
+        ],
+        locale: false,
+        permanent: false,
+        destination: "/fr/:path",
+      },
+      {
+        source: "/fr/:path",
+        has: [
+          {
+            type: "cookie",
+            key: "lang",
+            value: "en",
+          },
+        ],
+        locale: false,
+        permanent: false,
+        destination: "/en/:path",
+      },
+      {
+        source: "/fr/home",
+        locale: false,
+        permanent: false,
+        destination: "/fr/accueil",
+      },
+      {
+        source: "/fr/about",
+        locale: false,
+        permanent: false,
+        destination: "/fr/a-propos",
+      },
+      {
+        source: "/fr/signup",
+        locale: false,
+        permanent: false,
+        destination: "/fr/inscription",
+      },
+      {
+        source: "/fr/projects",
+        locale: false,
+        permanent: false,
+        destination: "/fr/projets",
+      },
+      {
+        source: "/fr/signup/privacy",
+        locale: false,
+        permanent: false,
+        destination: "fr/inscription/protection-renseignements-personnels",
+      },
+      {
+        source: "/fr/thankyou",
+        locale: false,
+        permanent: false,
+        destination: "/fr/merci",
+      },
+      {
+        source: "/fr/unsubscribe",
+        locale: false,
+        permanent: false,
+        destination: "/fr/desabonnement",
+      },
+    ];
+  },
 };